Add the Deathword
Up some of the drops that should be able to drop more than just 1

+++ src/main/java/gaia/entity/prop/AntHill.java
@@ -82,7 +82,7 @@ public class AntHill extends AbstractPropEntity {
 					++spawnTime;
 				} else {
 					if (!level.isClientSide) {
-						setSpawn((byte) 0);
+						setSpawn(0);
 					}
 
 					hurt(DamageSource.GENERIC, getMaxHealth() / getSpawnAmount());
